refact: Clock.ts 수정 및 위치 이동
